+++ tests/TestCase/Controller/Admin/VideosControllerTest.php
@@ -46,6 +46,11 @@ class VideosControllerTest extends IntegrationTestCase
     protected $Videos;
 
     /**
+     * @var array
+     */
+    protected $example;
+
+    /**
      * Fixtures
      * @var array
      */
@@ -71,6 +76,13 @@ class VideosControllerTest extends IntegrationTestCase
         parent::setUp();
 
         $this->setUserGroup('admin');
+
+        $this->example = [
+            'user_id' => 1,
+            'category_id' => 1,
+            'title' => 'new video title',
+            'text' => 'new video text',
+        ];
 
         $this->Controller = new VideosController;
 
@@ -192,5 +204,199 @@ class VideosControllerTest extends IntegrationTestCase
             3 => false,
             4 => true,
         ]);
+    }
+
+    /**
+     * Tests for `index()` method
+     * @test
+     */
+    public function testIndex()
+    {
+        $this->get(array_merge($this->url, ['action' => 'index']));
+        $this->assertResponseOk();
+        $this->assertResponseNotEmpty();
+        $this->assertTemplate(ROOT . 'src/Template/Admin/Videos/index.ctp');
+
+        $videosFromView = $this->viewVariable('videos');
+        $this->assertInstanceof('Cake\ORM\ResultSet', $videosFromView);
+        $this->assertNotEmpty($videosFromView);
+
+        foreach ($videosFromView as $video) {
+            $this->assertInstanceof('MeCmsYoutube\Model\Entity\Video', $video);
+        }
+    }
+
+    /**
+     * Tests for `add()` method
+     * @test
+     */
+    public function testAdd()
+    {
+        $url = array_merge($this->url, ['action' => 'add']);
+
+        $this->get($url);
+        $this->assertResponseOk();
+        $this->assertResponseNotEmpty();
+        $this->assertTemplate(ROOT . 'src/Template/Admin/Videos/add.ctp');
+
+        $url = array_merge($url, ['?' => ['url' => 'invalidUrl']]);
+
+        //GET request. Invalid url
+        $this->get($url);
+        $this->assertRedirect([]);
+        $this->assertSession('This is not a YouTube video', 'Flash.flash.0.message');
+
+        $url = array_merge($url, ['?' => ['url' => 'https://www.youtube.com/watch?v=aaa']]);
+
+        //GET request. Invalid Youtube ID
+        $this->get($url);
+        $this->assertRedirect([]);
+        $this->assertSession('Unable to retrieve video informations. Probably the video is private', 'Flash.flash.0.message');
+
+        $url = array_merge($url, ['?' => ['url' => 'https://www.youtube.com/watch?v=6z4KK7RWjmk']]);
+
+        //GET request. Now the url and Youtube ID are valid
+        $this->get($url);
+        $this->assertResponseOk();
+        $this->assertResponseNotEmpty();
+
+        $videoFromView = $this->viewVariable('video');
+        $this->assertInstanceof('MeCmsYoutube\Model\Entity\Video', $videoFromView);
+        $this->assertNotEmpty($videoFromView);
+
+        //POST request. Data are valid
+        $this->post($url, $this->example);
+        $this->assertRedirect(['action' => 'index']);
+        $this->assertSession('The operation has been performed correctly', 'Flash.flash.0.message');
+
+        //POST request. Data are invalid
+        $this->post($url, ['title' => 'aa']);
+        $this->assertResponseOk();
+        $this->assertResponseNotEmpty();
+        $this->assertResponseContains('The operation has not been performed correctly');
+
+        $videoFromView = $this->viewVariable('video');
+        $this->assertInstanceof('MeCmsYoutube\Model\Entity\Video', $videoFromView);
+        $this->assertNotEmpty($videoFromView);
+    }
+
+    /**
+     * Tests for `edit()` method
+     * @test
+     */
+    public function testEdit()
+    {
+        $url = array_merge($this->url, ['action' => 'edit', 1]);
+
+        $this->get($url);
+        $this->assertResponseOk();
+        $this->assertResponseNotEmpty();
+        $this->assertTemplate(ROOT . 'src/Template/Admin/Videos/edit.ctp');
+
+        $videoFromView = $this->viewVariable('video');
+        $this->assertInstanceof('MeCmsYoutube\Model\Entity\Video', $videoFromView);
+        $this->assertNotEmpty($videoFromView);
+
+        //Checks if the `created` field has been properly formatted
+        $this->assertRegExp('/^\d{4}\-\d{2}\-\d{2}\s\d{2}\:\d{2}$/', $videoFromView->created);
+
+        //POST request. Data are valid
+        $this->post($url, ['title' => 'another title']);
+        $this->assertRedirect(['action' => 'index']);
+        $this->assertSession('The operation has been performed correctly', 'Flash.flash.0.message');
+
+        //POST request. Data are invalid
+        $this->post($url, ['title' => 'aa']);
+        $this->assertResponseOk();
+        $this->assertResponseNotEmpty();
+        $this->assertResponseContains('The operation has not been performed correctly');
+
+        $videoFromView = $this->viewVariable('video');
+        $this->assertInstanceof('MeCmsYoutube\Model\Entity\Video', $videoFromView);
+        $this->assertNotEmpty($videoFromView);
+    }
+
+    /**
+     * Tests for `delete()` method
+     * @test
+     */
+    public function testDelete()
+    {
+        $this->post(array_merge($this->url, ['action' => 'delete', 1]));
+        $this->assertRedirect(['action' => 'index']);
+        $this->assertSession('The operation has been performed correctly', 'Flash.flash.0.message');
+    }
+
+    /**
+     * Tests that the admins and manangers can add and edit as another user
+     * @test
+     */
+    public function testAdminsAndManagersCanAddAndEditAsAnotherUser()
+    {
+        foreach (['admin', 'manager'] as $userGroup) {
+            $this->setUserGroup($userGroup);
+
+            foreach ([1, 2] as $userId) {
+                //Adds record
+                $this->post(
+                    array_merge($this->url, ['action' => 'add', '?' => ['url' => 'https://www.youtube.com/watch?v=6z4KK7RWjmk']]),
+                    array_merge($this->example, ['user_id' => $userId])
+                );
+                $this->assertRedirect(['action' => 'index']);
+                $this->assertSession('The operation has been performed correctly', 'Flash.flash.0.message');
+
+                $video = $this->Videos->find()->last();
+                $this->assertEquals($userId, $video->user_id);
+
+                //Edits record, adding +1 to the `user_id`
+                $this->post(
+                    array_merge($this->url, ['action' => 'edit', $video->id]),
+                    array_merge($this->example, ['user_id' => $userId + 1])
+                );
+                $this->assertRedirect(['action' => 'index']);
+                $this->assertSession('The operation has been performed correctly', 'Flash.flash.0.message');
+
+                $video = $this->Videos->findById($video->id)->first();
+                $this->assertEquals($userId + 1, $video->user_id);
+
+                $this->Videos->delete($video);
+            }
+        }
+    }
+
+    /**
+     * Tests that the other users cannot add and edit as another user
+     * @test
+     */
+    public function testOtherUsersCannotAddOrEditAsAnotherUser()
+    {
+        $this->setUserGroup('user');
+        $this->setUserId(3);
+
+        foreach ([1, 2] as $userId) {
+            //Adds record
+            $this->post(
+                array_merge($this->url, ['action' => 'add', '?' => ['url' => 'https://www.youtube.com/watch?v=6z4KK7RWjmk']]),
+                array_merge($this->example, ['user_id' => $userId])
+            );
+            $this->assertRedirect(['action' => 'index']);
+            $this->assertSession('The operation has been performed correctly', 'Flash.flash.0.message');
+
+            $video = $this->Videos->find()->last();
+            $this->assertEquals(3, $video->user_id);
+
+            //Edits record, adding +1 to the `user_id`
+            $this->post(
+                array_merge($this->url, ['action' => 'edit', $video->id]),
+                array_merge($this->example, ['user_id' => $userId + 1])
+            );
+            $this->assertRedirect(['action' => 'index']);
+            $this->assertSession('The operation has been performed correctly', 'Flash.flash.0.message');
+
+            $video = $this->Videos->findById($video->id)->first();
+            $this->assertEquals(3, $video->user_id);
+
+            $this->Videos->delete($video);
+        }
     }
 }